炮射击完成 2016-8-2-1:07

+++ driver/inc/Driver_FricMotor.h
@@ -37,7 +37,7 @@ void GunFric_Control(uint8_t Control);
 
 //小英雄枪炮一体增加炮摩擦轮
 #if INFANTRY == 6
-void FricArtillerySpeed_Adjust(uint8_t TargetSpeed);
+void FricArtillerySpeed_Adjust(int16_t TargetSpeed);
 void FricArtilleryMotorCurrent(int16_t LeftArtilleryCurrent, int16_t RightArtilleryCurrent);
 #endif
